Trait in_place_fastx::block::AbcProducer

pub trait AbcProducer {
    fn correct_block_size(block: &[u8]) -> Result<u64>;

    fn offset(&self) -> u64;

    fn file_length(&self) -> u64;

    fn file(&self) -> &File;

    fn blocksize(&self) -> u64;

    fn set_offset(&mut self, value: u64);

    fn next_block(&mut self) -> Result<Option<Block>> { ... }

    fn filesize<P>(path: &P) -> Result<u64>
    where
        P: AsRef<Path>,
    { ... }

    fn fix_blocksize<P>(path: &P, blocksize: u64) -> Result<u64>
    where
        P: AsRef<Path>,
    { ... }
}

Trait to produce block

Required methods

fn correct_block_size(block: &[u8]) -> Result<u64>

Search the begin of the partial record at the end of Block

fn offset(&self) -> u64

Get current value of offset

fn file_length(&self) -> u64

Get file length

fn file(&self) -> &File

Get file

fn blocksize(&self) -> u64

Get blocksize

fn set_offset(&mut self, value: u64)

Set value of offset

Provided methods

fn next_block(&mut self) -> Result<Option<Block>>

Get the next Block, all Block contains almost one record

fn filesize<P>(path: &P) -> Result<u64> where
P: AsRef<Path>,

Get file size

fn fix_blocksize<P>(path: &P, blocksize: u64) -> Result<u64> where
P: AsRef<Path>,

Fix blocksize