Trait in_place_fastx::parser::sequential::Sequential

pub trait Sequential<B, R> where
    B: Producer,
    R: Reader, {
    fn record(&mut self, _record: Record<'_>);

    fn parse<P>(&mut self, path: P) -> Result<()>
    where
        P: AsRef<Path>,
    { ... }

    fn with_blocksize<P>(&mut self, blocksize: u64, path: P) -> Result<()>
    where
        P: AsRef<Path>,
    { ... }

    fn block(&mut self, block: Block) -> Result<()> { ... }
}

Trait allow sequential parsing of fastq

Reading is perform by block. Parser map a block of file in memory, this block is resize to remove incomplete record. For each block record position is extract and record function is call on it.

Required methods

fn record(&mut self, _record: Record<'_>)

Method call to parse a record

Provided methods

fn parse<P>(&mut self, path: P) -> Result<()> where
P: AsRef<Path>,

Parse file indicate by path with default blocksize crate::DEFAULT_BLOCKSIZE

fn with_blocksize<P>(&mut self, blocksize: u64, path: P) -> Result<()> where
P: AsRef<Path>,

Parse file indicate by path with selected blocksize

fn block(&mut self, block: Block) -> Result<()>

Method call to parse a block